First commit
Single elm file implementing the basic model of series of slits and the
elm architecture requirements for drawing the slits and dragging slits up
and down within the array. The position of the dragged slit is checked to
prevent it overlapping with other slits or falling off either end of the array.

Code to allow resizing of slits (as opposed to moving them) is also largely in
place though not complete. The 'dragType' function just needs to be implemented
to detect whether the cursor is closer to the middle, top or bottom of the slit
when a DragStart message is received.

There seems to be a bug where X-Windows takes over and drags the slit
as an SVG object for a bit (X cursor shows) and the app loses events
causing the slit to become stuck in dragging mode even with the mouse
up. Clicking again stops it though. Not tested off Linux.