Use y-coordinates relative to svg to allow embedding in a web page
Elm's mouse events give the position within the whole page (they use the 'pageX'
and 'pageY' properties of the event), but the coordinates used in svg are
relative to the svg element. So the existing code was fine as long as the svg
was positioned in the top left of the window as the two coordinate systems are
the same. Embedding the app in a web page breaks this.

To fix this, DragStart events now also contain the 'offsetY' value, which is
relative to a transparent overlay of the slits graphic. Subtracting this from
the 'pageX' value gives the y-origin of the svg in page coords. This is stored
in the drag information in the model to allow the relative values to be
calculated for further elm mouse events while the slit is being dragged.